# src/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Constellate API",
		"version": "1.0.0",
		"description": "Federated event management platform implementing ActivityPub protocol",
		"license": {
			"name": "AGPL-3.0"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Development server"
		}
	],
...
	"components": {
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "better-auth.session_token",
				"description": "Session-based authentication using cookies"
			},
			"apiKeyAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "API Key",
				"description": "API key authentication. Use format: Bearer sk_live_..."
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string",
						"description": "Error code (e.g., VALIDATION_ERROR, UNAUTHORIZED, etc.)"
					},
					"message": {
						"type": "string",
						"description": "Human-readable error message"
					},
					"details": {
						"description": "Additional error details (only included in development mode)",
						"oneOf": [
							{
								"type": "object"
							},
							{
								"type": "array"
							}
						]
					}
				},
				"required": ["error", "message"]
			},
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"displayColor": {
						"type": "string",
						"nullable": true
					},
					"profileImage": {
						"type": "string",
						"nullable": true
					},
					"isPublicProfile": {
						"type": "boolean",
						"description": "Profile visibility: true = public (visible to everyone), false = private (only visible to followers). Defaults to true.",
						"default": true
					}
				}
			},
...
			"NotificationListResponse": {
				"type": "object",
				"properties": {
					"notifications": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Notification"
						}
					},
					"unreadCount": {
						"type": "integer"
					}
				}
			},
...
		}
	},
	"security": [
		{
			"cookieAuth": []
		},
		{
			"apiKeyAuth": []
		}
	],
	"paths": {
		"/api/notifications": {
			"get": {
				"tags": ["Notifications"],
				"summary": "List notifications for current user",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20,
							"minimum": 1,
							"maximum": 100
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of notifications",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
...
		"/api/events": {
			"post": {
				"tags": ["Events"],
				"summary": "Create a new event",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EventInput"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Event created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Event"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"400": {
						"description": "Validation error - occurs when input validation fails (e.g., invalid tags, missing required fields, or tag normalization fails)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								},
								"examples": {
									"tagValidationError": {
										"summary": "Tag validation error",
										"value": {
											"error": "VALIDATION_ERROR",
											"message": "Failed to process tags: ensure tags are valid strings"
										}
									},
									"schemaValidationError": {
										"summary": "Schema validation error",
										"value": {
											"error": "VALIDATION_ERROR",
											"message": "Invalid input data",
											"details": [
												{
													"code": "invalid_type",
													"expected": "string",
													"received": "number",
													"path": ["title"],
													"message": "Expected string, received number"
												}
											]
										}
									}
								}
							}
						}
					}
				}
			},
			"get": {
				"tags": ["Events"],
				"summary": "List events",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of events",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"events": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Event"
											}
										},
										"pagination": {
											"type": "object",
											"properties": {
												"page": {
													"type": "integer"
												},
												"limit": {
													"type": "integer"
												},
												"total": {
													"type": "integer"
												},
												"pages": {
													"type": "integer"
												}
											}
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/search": {
			"get": {
				"tags": ["Search"],
				"summary": "Search events",
				"parameters": [
					{
						"name": "q",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Search query (searches in title and summary)"
					},
					{
						"name": "location",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter by location"
					},
					{
						"name": "startDate",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "date-time"
						},
						"description": "Filter events starting from this date"
					},
					{
						"name": "endDate",
						"in": "query",
						"schema": {
							"type": "string",
							"format": "date-time"
						},
						"description": "Filter events ending before this date"
					},
					{
						"name": "status",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["EventScheduled", "EventCancelled", "EventPostponed"]
						},
						"description": "Filter by event status"
					},
					{
						"name": "mode",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": [
								"OfflineEventAttendanceMode",
								"OnlineEventAttendanceMode",
								"MixedEventAttendanceMode"
							]
						},
						"description": "Filter by attendance mode"
					},
					{
						"name": "username",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Filter by organizer username"
					},
					{
						"name": "tags",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Comma-separated list of tags to filter by"
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20,
							"maximum": 100
						}
					}
				],
				"responses": {
					"200": {
						"description": "Search results",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Event"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/recommendations": {
			"get": {
				"tags": ["Recommendations"],
				"summary": "Get personalized event recommendations",
				"description": "Returns upcoming events scored for the authenticated user based on recent attendance, likes, followed organizers, and current engagement.",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "Maximum number of recommendations to return (1-20)",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 20,
							"default": 10
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of recommended events",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"recommendations": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"event": {
														"$ref": "#/components/schemas/Event"
													},
													"score": {
														"type": "number",
														"description": "Composite recommendation score"
													},
													"reasons": {
														"type": "array",
														"items": {
															"type": "string"
														}
													},
													"signals": {
														"type": "object",
														"properties": {
															"matchedTags": {
																"type": "array",
																"items": {
																	"type": "string"
																}
															},
															"followedOrganizer": {
																"type": "boolean"
															},
															"hostAffinity": {
																"type": "boolean"
															},
															"popularityScore": {
																"type": "number"
															}
														},
														"required": [
															"matchedTags",
															"followedOrganizer",
															"hostAffinity",
															"popularityScore"
														]
													}
												},
												"required": ["event", "score", "reasons", "signals"]
											}
										},
										"metadata": {
											"type": "object",
											"properties": {
												"generatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"signals": {
													"type": "object",
													"properties": {
														"tags": {
															"type": "integer"
														},
														"hosts": {
															"type": "integer"
														},
														"followed": {
															"type": "integer"
														}
													}
												}
											}
										}
									},
									"required": ["recommendations"]
								}
							}
						}
					},
					"401": {
						"description": "Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookieAuth": []
					}
				]
			}
		},
...
		"/api/location/search": {
			"get": {
				"tags": ["Location"],
				"summary": "Search for a place and return coordinates",
				"parameters": [
					{
						"name": "q",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 3
						},
						"description": "Free-form place query"
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 10,
							"default": 5
						},
						"description": "Maximum suggestions to return"
					}
				],
				"responses": {
					"200": {
						"description": "Location suggestions",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"query": {
											"type": "string"
										},
										"results": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/LocationSuggestion"
											}
										}
									},
									"required": ["query", "results"]
								}
							}
						}
					}
				}
			}
		},
		"/api/user-search": {
			"get": {
				"tags": ["User Search"],
				"summary": "Search for users (WebFinger)",
				"parameters": [
					{
						"name": "query",
						"in": "query",
						"required": true,
						"description": "Username or @user@domain format",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Search results",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/User"
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/api/admin/users": {
			"get": {
				"tags": ["Admin"],
				"summary": "List users",
				"description": "List all users with pagination and filtering. Admin only.",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1,
							"minimum": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20,
							"minimum": 1,
							"maximum": 100
						}
					},
					{
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Search by username, name, or email"
					},
					{
						"name": "isBot",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["true", "false"]
						},
						"description": "Filter by bot status"
					}
				],
				"responses": {
					"200": {
						"description": "User list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden (not admin)"
					}
				}
			},
			"post": {
				"tags": ["Admin"],
				"summary": "Create user",
				"description": "Create a new user. For bot users, password is not required. For regular users, password is required. Admin only.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserInput"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminUser"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or username/email already exists"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden (not admin)"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
		"/api/setup": {
			"post": {
				"tags": ["Setup"],
				"summary": "Create first admin user",
				"description": "Creates the initial admin user for a new instance. Only works when no users exist.",
				"security": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["email", "username", "name"],
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"description": "Admin user email address"
									},
									"username": {
										"type": "string",
										"minLength": 1,
										"maxLength": 50,
										"description": "Admin username"
									},
									"name": {
										"type": "string",
										"description": "Admin display name"
									},
									"password": {
										"type": "string",
										"minLength": 8,
										"description": "Optional password. If not provided, user can only login via magic link."
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Admin user created successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean"
										},
										"user": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"username": {
													"type": "string"
												},
												"name": {
													"type": "string"
												}
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Missing required fields",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"403": {
						"description": "Setup already completed (users exist)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"description": "Setup failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/api/instances": {
			"get": {
				"tags": ["Instances"],
				"summary": "List known federated instances",
				"description": "Returns a list of federated instances discovered through ActivityPub activities, with statistics about connections and activity",
				"security": [
					{
						"cookieAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"description": "Maximum number of instances to return (max 100)",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 100,
							"default": 50
						}
					},
					{
						"name": "offset",
						"in": "query",
						"description": "Number of instances to skip for pagination",
						"schema": {
							"type": "integer",
							"minimum": 0,
							"default": 0
						}
					},
					{
						"name": "sortBy",
						"in": "query",
						"description": "Sort instances by activity, user count, or creation date",
						"schema": {
							"type": "string",
							"enum": ["activity", "users", "created"],
							"default": "activity"
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of federated instances",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"instances": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": { "type": "string" },
													"domain": { "type": "string" },
													"baseUrl": { "type": "string" },
													"software": {
														"type": "string",
														"nullable": true
													},
													"version": {
														"type": "string",
														"nullable": true
													},
													"title": { "type": "string", "nullable": true },
													"description": {
														"type": "string",
														"nullable": true
													},
													"iconUrl": {
														"type": "string",
														"nullable": true
													},
													"userCount": {
														"type": "integer",
														"nullable": true
													},
													"eventCount": {
														"type": "integer",
														"nullable": true
													},
													"lastActivityAt": {
														"type": "string",
														"format": "date-time",
														"nullable": true
													},
													"isBlocked": { "type": "boolean" },
													"stats": {
														"type": "object",
														"properties": {
															"remoteUsers": { "type": "integer" },
															"remoteEvents": { "type": "integer" },
															"localFollowing": { "type": "integer" }
														}
													}
												}
											}
										},
										"total": { "type": "integer" },
										"limit": { "type": "integer" },
										"offset": { "type": "integer" }
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/instances/search": {
			"get": {
				"tags": ["Instances"],
				"summary": "Search for instances",
				"description": "Search known instances by domain name, title, or description",
				"security": [
					{
						"cookieAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "q",
						"in": "query",
						"required": true,
						"description": "Search query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"description": "Maximum number of results (max 50)",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 50,
							"default": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "Search results",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"instances": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": { "type": "string" },
													"domain": { "type": "string" },
													"baseUrl": { "type": "string" },
													"title": { "type": "string", "nullable": true },
													"description": {
														"type": "string",
														"nullable": true
													},
													"software": {
														"type": "string",
														"nullable": true
													},
													"version": {
														"type": "string",
														"nullable": true
													},
													"isBlocked": { "type": "boolean" }
												}
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid search query"
					}
				}
			}
		},
...
	}
}